--- keyboards/rgbkb/mun/mun.h
@@ -1,16 +1,0 @@
-/*
- * ----------------------------------------------------------------------------
- * "THE BEER-WARE LICENSE" (Revision 42):
- * <https://github.com/Legonut> wrote this file.  As long as you retain this
- * notice you can do whatever you want with this stuff. If we meet some day, and
- * you think this stuff is worth it, you can buy me a beer in return. David Rauseo
- * ----------------------------------------------------------------------------
- */
-
-#pragma once
-
-#if defined(KEYBOARD_rgbkb_mun_rev1)
-#    include "rev1.h"
-#endif
-
-#include "quantum.h"
