+++ keyboards/rgbkb/mun/rev1/keyboard.json
@@ -8,6 +8,19 @@
         "pid": "0x3505",
         "device_version": "0.0.1"
     },
+    "build": {
+        "lto": true
+    },
+    "features": {
+        "bootmagic": false,
+        "encoder": true,
+        "extrakey": true,
+        "mousekey": false,
+        "nkro": true,
+        "oled": true,
+        "rgb_matrix": true,
+        "rgblight": false
+    },
     "rgblight": {
         "led_count": 98,
         "split_count": [49, 49],
