+++ keyboards/rgbkb/mun/rev1/rev1.h
@@ -9,5 +9,5 @@
 
 #pragma once
 
-#include "mun.h"
+#include "quantum.h"
 #include "touch_encoder.h"
